refactor(example): use barrel import for book module
#1684
