+++ projects/example-app/src/app/books/actions/book.actions.ts
@@ -1,5 +1,5 @@
+import { Book } from '@example-app/books/models';
 import { createAction, props } from '@ngrx/store';
-import { Book } from '@example-app/books/models/book';
 
 export const loadBook = createAction(
   '[Book Exists Guard] Load Book',
