+++ projects/example-app/src/app/books/components/book-authors.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
 
-import { Book } from '@example-app/books/models/book';
+import { Book } from '@example-app/books/models';
 
 @Component({
   selector: 'bc-book-authors',
