+++ projects/example-app/src/app/books/containers/find-book-page.component.ts
@@ -1,11 +1,11 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+import { FindBookPageActions } from '@example-app/books/actions';
+import { Book } from '@example-app/books/models';
+import * as fromBooks from '@example-app/books/reducers';
 import { select, Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
 import { take } from 'rxjs/operators';
-
-import { FindBookPageActions } from '@example-app/books/actions';
-import { Book } from '@example-app/books/models/book';
-import * as fromBooks from '@example-app/books/reducers';
 
 @Component({
   selector: 'bc-find-book-page',
