+++ projects/example-app/src/app/books/containers/index.ts
@@ -1,0 +1,4 @@
+export * from './collection-page.component';
+export * from './find-book-page.component';
+export * from './selected-book-page.component';
+export * from './view-book-page.component';
