+++ projects/example-app/src/app/books/effects/collection.effects.ts
@@ -1,14 +1,15 @@
 import { Injectable } from '@angular/core';
-import { Actions, Effect, ofType, createEffect } from '@ngrx/effects';
-import { defer, of } from 'rxjs';
-import { catchError, map, mergeMap, switchMap } from 'rxjs/operators';
-import { Book } from '@example-app/books/models/book';
 import {
   CollectionApiActions,
   CollectionPageActions,
   SelectedBookPageActions,
 } from '@example-app/books/actions';
+import { Book } from '@example-app/books/models';
 import { BookStorageService } from '@example-app/core/services';
+import { Actions, createEffect, ofType } from '@ngrx/effects';
+import { defer, of } from 'rxjs';
+import { catchError, map, mergeMap, switchMap } from 'rxjs/operators';
+
 @Injectable()
 export class CollectionEffects {
   /**
