+++ projects/example-app/src/app/core/services/book-storage.service.spec.ts
@@ -1,10 +1,10 @@
 import { TestBed } from '@angular/core/testing';
-import { cold } from 'jasmine-marbles';
-import { Book } from '@example-app/books/models/book';
+import { Book } from '@example-app/books/models';
 import {
   BookStorageService,
   LOCAL_STORAGE_TOKEN,
 } from '@example-app/core/services/book-storage.service';
+import { cold } from 'jasmine-marbles';
 
 describe('BookStorageService', () => {
   let fixture: any;
